[6/6] [nexus] convert external API to trait (#6373)
At long last, here's the last OpenAPI document converted into a trait: the
Nexus external API. This is a fairly large change, but with all the
prerequisites in place the conversion is overall pretty mechanical. I did need
to move some types around, but the bulk of the PR is just adding the very large
external API trait.

This PR also adds a `.git-blame-ignore-revs` entry corresponding to #6372, so that the indent doesn't ruin blame in the GitHub web view. (You can configure this locally as well with `git config blame.ignoreRevsFile .git-blame-ignore-revs`.)

In a subsequent PR I'll remove the old instructions.

Depends on:

* #6355
* #6364 
* #6369
* #6370
* #6372

Co-authored-by: David Crespo <david.crespo@oxidecomputer.com>