add `VmmState::Creating`
This commit introduces a new `VmmState` variant, `Creating`. The purpose
of this variant is to differentiate between VMMs which have been created
in the database but do not yet exist on a sled, from those which are
known to sled-agent and are actually starting/migrating.

This is necessary because presently, the `instance-watcher` background
task will attempt to perform health checks for `Starting` and
`Migrating` VMMs, and --- if the sled-agent does not actually know about
the VMM yet --- will move it to `Failed`. But, when a VMM is newly
created, the sled-agent won't know about it yet until it has been
registered with the sled-agent, so if the `instance-watcher` task
activates while an `instance-start` or `instance-migrate` saga is in
flight, the `instance-watcher`` task may incorrectly move its VMM to
`Failed` if the VMM record has been created but not yet registered. Now,
after adding the `Creating` variant, we can avoid performing health
checks for those VMMs until the sled-agent actually acknowledges a
request to register the VMM.