Skip `test_disk_usage` if there's no `/usr/bin/du`
The `zone_bundle::disk_usage` function in sled-agent executes `du(1)` to
measure disk usage for a path. Presently, the path to the `du(1)`
executable is hard-coded to `/usr/bin/du`. On some systems, such as
NixOS, the `du(1)` executable may not live at the path `/usr/bin/du`. We
could, potentially, change the `disk_usage` function to find a `du(1)`
binary on the PATH, or fall back to other locations...but, on Helios, it
is always at `/usr/bin/du`, so hardcoding it is a good solution for the
environment that sled-agents are actually intended to run in.

However, hardcoding the path means that this test will always fail on
systems where there is no `/usr/bin/du`. Therefore, this commit changes
the test to first check if `/usr/bin/du` exists, and return early if it
does not, printing a warning indicating that the test is skipped.

If, someday, Rust's libtest grows a mechanism for tests to mark
themselves as skipped at runtime dynamically (rather than using the
`#[ignore]` attribute at compile-time), we should probably switch to
using that hypothetical mechanism here. See this
[internals.rust-lang.org thread][1].

If folks aren't happy with making the test skip itself, we *could*
change the implementation to worth with `du(1)` being at a different
location by finding it from the PATH, potentially falling back to
`/usr/bin/du` if it isn't on the PATH. My thought process was that it
was better to avoid complicating the production sled-agent code just to
support environments that the software will never actually run on for
test purposes, but I'm open to being convinced otherwise.

[1]: https://internals.rust-lang.org/t/pre-rfc-skippable-tests/14611