remove `patch.crates-io.zone` (#6484)
While looking at doing lockfile maintenance for R11 I noticed that we
patch the `zone` package to be a branch that adds Eq and Hash to
`zone::State`. This branch was never merged into the main branch of
oxidecomputer/zone, and also appears to be unnecessary for Omicron to
build at present. This switches us back to the released version on
crates.io.