update to track `SledFailuresOnly` policy
2db6eff96c0b2d7f94999b2a6332908fce09aa7e added a `SledFailuresOnly`
auto-restart policy in addition to `Never` and `AllFailures`. I
discussed the rationale for that in [this comment][1]. Currently, there
isn't a mechanism to detect whether an instance is `Failed` because the
individual instance crashed or because the whole sled was restarted, so
for now, we assume all failures are instance-level. But, we still need
to handle the new variant.

[1]: https://github.com/oxidecomputer/omicron/pull/6499#issuecomment-2325131201