[nexus] Mark VMMs on `Expunged` sleds as `Failed`
When a sled is expunged, any VMMs previously believed to be on that sled
are definitely no longer present. Currently, expunging a sled will not
do anything to VMMs resident on that sled, and the instances will remain
in the `Running` state...but it won't be possible to stop them, as their
sled-agent will probably no longer be reachable.

This commit updates the `instance-watcher` background task to mark any
VMMs on `Expunged` sleds as `Failed`, so that those VMMs' instances can
transition to `Failed` and be restarted or deleted normally. It seems
better to do this in the `instance-watcher` background task, rather than
in the code paths for actually marking a sled as expunged, because
transitioning a VMM to failed triggers an instance-update saga to move
the instance to failed, and doing all of that in the transaction that
expunges the sled seems unfortunate --- expunging the sled would be a
much longer-running operation, and we'd have to be careful to not roll
it back if one update saga fails or whatever. So, instead, I think it's
nicer to just mark the sled as expunged and then let the background task
do the remaining work. And, the `instance-watcher` background task
already queries for all instances *and their associated sleds*, so it
makes sense to do the transition there rather than in the
`instance-updater` background task, since we'd have to add an additional
JOIN to look at sleds, which seems unfortunate. Also, this way,
VMMs that are marked as `Failed` due to sled expungement can also be
reflected in the metrics, which seems nice.

This was, overall, pretty straightforward: I've just removed the
`SledFilter::InService` from the `instance_and_vmm_list_by_sled_agent`
query,[^1], passed the `Sled` into the `check_instance` function, and
added a code path to skip the actual HTTP health check if the sled is
`Expunged` and just go ahead and mark the VMM as `Failed`. I think it
still makes sense to spawn a task for these VMMs, even though they don't
actually do HTTP requests to the sled-agent, because they will attempt
to run a whole `instance-update` saga to transition the instance to
`Failed`, which can take a bit of time. It seems nice to just keep
starting other health checks while that runs, instead of doing it
sequentially.

[^1]: Apparently there were actually *two* separate
    `SledFilter::InService` calls, which tripped  me up for longer than
    I would like to admit...