[nexus] Turn `instance.boot_on_fault` into an enum (#6499)
Currently, the `instance` table has a `boot_on_fault` column, which is a
`bool`. This is intended to indicate whether an instance should be
automatically restarted in response to failures, although currently, it
isn't actually consumed by anything. However, as determined in RFD 486,
Nexus will eventually grow functionality for automatically restarting
instances that have transitioned to the `Failed` state, if
`boot_on_fault` is set.

@gjcolombo suggests that this field should probably be replaced with an
enum, rather than a `bool`. This way, we could represent more
boot-on-fault disciplines than "always reboot on faults" and "never do
that". Since nothing is currently consuming this field, it's probably a
good idea to go ahead and do the requisite SQL surgery to turn it into
an enum _now_, before we write code that actually consumes
it...especially since I'm planning on actually doing that next (see
#6491).

This commit replaces the `boot_on_fault` column with a new
`auto_restart_policy` column. The type for this column is a newly-added
`InstanceAutoRestart` enum, which currently has variants for
`AllFailures` (the instance should be automatically restarted any time
it fails), `SledFailuresOnly`, (the instance should be restarted if the
sled it's on reboots or is expunged, but it should *not* be restarted if
its individual VMM process crashes), and `Never` (the instance should
never be automatically restarted). The database migration adding
`auto_restart_policy` backfills any existing `boot_on_fault: true`
instances with `InstanceAutoRestart::AllFailures` (although, because
users can't currently set a value for `boot_on_fault`, there should
usually be no such instances). For instances with `boot_on_fault:
false`, the auto-restart policy is left `NULL`; in the future, we can
determine what the default policy should be (perhaps at the
project-level) and backfill these `NULL`s as well. For now, instances
with an unset (`NULL`) policy will not be restarted.

Closes #6490