[db-queries] Allow join expressions in `paginated-multicolumn` (#6530)
Currently, the `paginated_multicolumn` utility in
`nexus_db_queries::pagination` only works when the select expression to
paginate is a table, and both columns to order by come from that table.
This means that it cannot easily be used to fix the bug in
`instance_and_vmm_list_by_sled_agent` that @davepacheco describes in
[this comment][1], which would require using `paginated_multicolumn` to
paginate on two columns in an inner join expression. 

This commit changes the giant wad of Diesel type ceremony on
`paginated_multicolumn` in order to ~~make it even worse~~ allow
expressions which are not tables to be paginated. I've added a test
demonstrating that this does, in fact, work.

Figuring out how to do this was...certainly an experience which I have
had. I think I need to lie down now.

[1]:
https://github.com/oxidecomputer/omicron/pull/6519#pullrequestreview-2283593739