SAML relay state shouldn't result in internal server error (#6537)
The relay state can be set by an IDP to some string value. Nexus itself
is expecting that string to be in the form of JSON that
serializes/deserializes to our `RelayState` type. With this fix we now
attempt to parse this value so that we know where to redirect the user
upon successful authentication, however if we can't we just send the
user to the root aka `/`.