Add retry loop to deleting timeseries by name (#6504)
We currently support deleting timeseries by name in a schema upgrade.
This deletion is implemented as a mutation in ClickHouse, which walks
all affected data parts and deletes the relevant records in a merge
operation. That's asynchronous by default, and run in a pool of
background tasks. Despite that, with large tables, it can take a while
for each mutation to complete, which blocks the server from queueing new
deletion requests. This can lead to timeouts, like seen in #6501.

This should fix #6501, but I'm not certain of that because I don't have
a good way to reproduce the bug. It seems likely that this is only seen
when the database is already heavily loaded, as it might be when doing
these mutations on large tables.