just pass the `InstanceEnsureBody` around
Passing the `InstanceEnsureBody` all the way through to the
instance-runner task, instead of unpacking it and passing its pieces as
arguments, lets us get rid  of the
`#[allow::clippy_too_many_arguments]` that are sprinkled around. It also
means that if a subsequent change adds additional fields to the
instance-ensure payload, we don't need to change all these function
signatures just to add them.