[nexus/sled-agent] Rip out `api::internal::InstanceRuntimeState` (#6556)
The `omicron_common::api::internal::InstanceRuntimeState` type was once
a key component of how Nexus and sled-agents communicated about
instances. Now, however, it's outlived its usefulness: #5749 changed
Nexus' `cpapi_instances_put` and sled-agent's `instance_get` APIs to
operate exclusively on `VmmRuntimeState`s rather than
`InstanceRuntimeState`, with the sled-agent almost completely unaware of
`InstanceRuntimeState`.

At present, the `InstanceRuntimeState` type remains in the internal API
just because it's used in the `InstanceEnsureBody` type, which is sent
to sled-agent's `instance_ensure_registered` API when a new VMM is
registered. However, looking at the code that
`instance_ensure_registered` calls into in sled-agent, we see that the
only thing from the `InstanceRuntimeState` that the sled-agent actually
*uses* is the migration ID of the current migration (if the instance is
migrating in). Soooo...we can just replace the whole
`InstanceRuntimeState` there with a `migration_id: Option<Uuid>`. Once
that's done, there are now no longer any internal APIs in Nexus or in
sled-agent that actually use `InstanceRuntimeState`, so the type can be
removed from the internal API entirely. All of the code in Nexus that
deals with instance runtime states already has changed to operate
exclusively on the `nexus_db_model` version of the struct, so removing
the API type is actually quite straightforward.

In addition, I've refactored the sled-agent `instance_ensure_registered`
code paths a little bit to pass around the `InstanceEnsureBody`, rather
than unpacking its fields at the HTTP entrypoint and then passing them
all through the chain of method calls that actually gets it to the
instance-runner task. Just passing the struct around means we can remove
a bunch of `#[allow(clippy::too_many_arguments)]` attributes. It also
makes changing the contents of the instance-ensure payload a bit easier,
as we'll no longer need to change the arguments list in the whole maze
of tiny little functions, all alike, that pass around all the pieces of
the `InstanceEnsureBody`.