Add next-item query based on a join (#6551)
- Adds the `NextItemSelfJoined` query, which implements the next-item
query as a self-join. The current implementation joins the target table
with the series of all possible values for the target item. Because CRDB
eagerly evaluates subqueries, this entire list must be buffered in
memory. The self-join version instead joins the existing table with the
literal next item (item + 1), and finds the lowest one that's not
allocated yet. Some initial testing suggests this can consume much less
memory and run much faster than the previous query implementations. As a
tradeoff, this new query cannot be used to randomly select _any_ item in
a range -- only the next, lowest value can be selected. Note that this
changes the way we allocate MAC addresses, which were previously random.
- Fixes #5904