Return Volume's own regions when soft-deleting (#6555)
Correctly populate the "regions" field in the section of the
DecreaseCrucibleResourceCountAndSoftDeleteVolume CTE that constructs the
V3 version of the CrucibleResources object. Previously this would almost
never return the Volume's regions, causing those to be picked up later
by `find_deleted_volume_regions` and cleaned up there.

This bug did not result in any leaked regions that this author knows
about, but did cause a lot of confusion when examining another unrelated
issue.