add concurrency limit, query order randomization
This commit reworks the instance-reincarnation task substantially to use
a randomized query order rather than pagination, as @gjcolombo described
in [this comment][1]. In addition, we now enforce a real limit on the
number of concurrently-running instance start sagas, which is presently
16, a number I picked completely arbitrarily.

This branch now depends on PR #6574, which --- it turns out --- is
actually a requirement for any kind of saga concurrency limiting in
background tasks that start a bunch of sagas. Whoopsie.

[1]: https://github.com/oxidecomputer/omicron/issues/6565#issuecomment-2347241408