remove failure reason check constraint
This breaks stuff like deleting a `Failed` instance moving its state
to `Destroyed` without clearing the last failure reason. Alternatively,
we could unset the failure reason when transitioning out of `Failed`,
but leaving it seems nicer. Since it's called "LAST failure reason", its
presence doesn't necessarily imply that the instance is currently
failed, and keeping the last failure reason when transitioning out of
`Failed` might be useful, both for UI purposes and for stuff like
checking if a new failure has the same reason as the prior one. We
*could* unset it, but that just means discarding information.

The check constraint should still apply for VMMs, but it's been changed
to allow VMMs with failure reasons to be Failed *or* Destroyed.