rework auto-restart cooldown
In order to surface these cooldowns in UIs, we need to be able to add
the cooldown period to the `external::Instance` type. Therefore, I've
changed it from being a config parameter to a constant defined on the
`ReincarnationFilter` type, so that we can fill in the field on the API
type. In the future I think we'll want to allow overriding this on the
project-level, and perhaps also the instance level.