Allow different versions of the RoT and bootloader
Currently we require all versions of the RoT and bootloader
for a particular type (sled, psc, switch) to have identical
versions. This is a stricter requirement than what we actually
care about: All images signed with the same key should be the
same version. Relax this requirement for both the RoT and
its bootloader