Attempt to handle "kstat-based metrics produce samples from the 1980's"
Closes https://github.com/oxidecomputer/omicron/issues/5899.

Instead of checking for dates or anything based on the data sampled, this
update starts collecting kstat samples (i.e. being interested in them) once
the sled agent is synchronized with NTP.

We leverage the metrics manager associated with the agent to now look
for a new message and update the tracked links we've added/tracked thus
far.