Add planner code for replicated clickhouse processes (#6566)
Primarily, this commit includes a planning mechanism to allocate
clickhouse keepers and servers to their in-service zones. This
`ClickhouseAllocator` is intended to be created by the
`BlueprintBuilder` after zones are added and expunged. It very carefully
manages adding or removing only one keeper node at a time and there are
a bunch of tests regarding this behavior.

In order to know when to continue allocating new keepers or servers or
expunging them, the allocator takes in the set of zones from the current
builder run, the parent configuration from the last blueprint, and any
inventory updates related to clickhouse keepers. We only pass in the
`latest_clickhouse_keeper_membership` from the inventory. This is all
that the allocator needs and also simplifies testing by precluding
having to build a complete `Collection` just to test. While we pass in
all `blueprint_zones` from the `BlueprintBuilder` to the
`ClickhouseAllocator::new` constructor, to make production code simpler
and safer to write, we mnaually manipulate the allocator to keep the
tests simple, and not have to worry about generating
`BlueprintZonesConfig` that we don't actually use.

None of this code is wired up to run, but all the planning logic should
be capable of being hooked in easily. We'll also need to add DB models
and queries, and write the executor, and inventory collection in follow
ups.