Abstract over kinds of ClickHouse deployments in tests (#6593)
- Add the `ClickHouseDeployment` enum, which manages an entire
ClickHouse deployment in test code, either a single-node for most tests,
or a cluster where relevant. For the cluster variant, this adds a way to
wait for the first child or all children, to be shutdown. This fixes a
bug in the logic for managing child processes, where failures of one of
the process could make zombies out of all the others. This also collects
the nodes into arrays, so we can resize the cluster easily if we want,
which fixes #4460.
- Use the new enum in the `ControlPlaneTestContext` for all Nexus
integration tests.
- Rework the `ch-dev` binary to use the new enum, and also print much
more verbose information about what it's doing when starting ClickHouse.
This fixes #3011.