Soft-delete volumes in a transaction, not CTE (#6623)
Change the process of soft-deleting a volume from a CTE into a
transaction. This is done to make changes easier to apply and review.

One part of the CTE that remains is the query fragment that
conditionally reduces volume references for region_snapshot rows, then
returns only the rows updated. It was easier to keep this than figure
out how to get it to work in diesel!