Talk to ClickHouse over the native protocol (#6584)
Adds a module in `oximeter_db` that speaks the native ClickHouse
protocol, a message-oriented protocol directly over TCP. This includes
the core messaging and serde needed for making a handshake with the
server, and then running most standard SQL queries.

This uses the codecs from `tokio_util` to turn the TCP stream into a
sink / stream of messges, which is much more convenient to operate on.
The client and server exchange a handshake, after which the client may
run SQL queries to select data.

There are a number of tests in the module, including low-level
serialization checks and asserting the results of actual SQL queries. In
addition, this adds a barebones SQL shell -- this is not intended to
replace the feature-rich official CLI, but is very useful for testing
the protocol implementation by running arbitrary queries.

This module is not consumed anywhere beyond the shell itself. A
follow-up commit will integrate it into the existing
`oximeter_db::Client` object, with the primary goal of making OxQL
queries much more flexible and efficient.